+++ NEWS.md
@@ -11,6 +11,7 @@
 * `tm_variable_browser` now display character variable summary statistics table in decreasing order of occurrences instead of alphabetical. 
 * `tm_variable_browser` displays histograms and statistics table for variables of type `Date`, `POSIXct` and `POSIXlt`.
 * Remove keys from missing data combinations plot in `tm_missing_data` when they have not been selected. 
+* Add rug plot to both axis option in `tm_g_scatterplot`.
 * issues with no news:
 
 # teal.modules.general 0.2.6